WiFi_Sniffer_Serial_Target/requirements.txt
keyboard==0.13.4
pyserial==3.4
tqdm==4.41.0
